requirements.txt
discord.py[voice]
flask
motor
gTTS
